feat: support updating an existing comment instead of creating one

+++ src/index.js
@@ -14,6 +14,7 @@ const {
 	PR_LABELS,
 	CREATE_COMMENT,
 	DELETE_EXISTING_COMMENT,
+	UPDATE_EXISTING_COMMENT,
 	PR_PREVIEW_DOMAIN,
 	ALIAS_DOMAINS,
 	ATTACH_COMMIT_METADATA,
@@ -110,7 +111,7 @@ const run = async () => {
 		}
 
 		if (IS_PR) {
-			if (DELETE_EXISTING_COMMENT) {
+			if (DELETE_EXISTING_COMMENT && !UPDATE_EXISTING_COMMENT) {
 				core.info('Checking for existing comment on PR')
 				const deletedCommentId = await github.deleteExistingComment()
 
@@ -140,8 +141,8 @@ const run = async () => {
 					[View Workflow Logs](${ LOG_URL })
 				`
 
-				const comment = await github.createComment(body)
-				core.info(`Comment created: ${ comment.html_url }`)
+				const comment = await github.createComment(body, UPDATE_EXISTING_COMMENT)
+				core.info(`Commented: ${ comment.html_url }`)
 			}
 
 			if (PR_LABELS) {
